Remove Dylib file, implement code branch directly into geGL
