+++ src/geGL/geGL.cpp
@@ -56,10 +56,53 @@ namespace ge{
         }
     };
 #else
-      
-#include <geGL/Dylib.h>
-      
+#include <dlfcn.h>
 #if __APPLE__
+      /** Wraps function dlopen() and print dlerror() message if error occured. */
+      void* dl_open_lib(char const *lib_name) {
+          void* lib = dlopen(lib_name, RTLD_LAZY);
+          if (lib == NULL)
+              fprintf(stderr, "dlopen(): %s\n", dlerror());
+          return lib;
+      }
+      
+      /** Wraps function dlclose() and print dlerror() message if error occured. */
+      int dl_close_lib(void *lib) {
+          int code = dlclose(lib);
+          if (code)
+              fprintf(stderr, "dlclose(): %s\n", dlerror());
+          return code;
+      }
+      
+      /* About macOS frameworks:
+       Some dynamic libraries are packed into frameworks in macOS. Function dlopen() is able to detect and search
+       for frameworks if input path looks like framework path (e.g. /stuff/Foo.framework/Foo). Load them (or only
+       it's part by adding additional subpath) by passing full framework name in format 'Foo.framework/Foo' for
+       framework named 'Foo'. Default fallback search paths (if DYLD_FALLBACK_FRAMEWORK_PATH was not set) ensure
+       that framework is searched in usual directories (for more information see '$ man dlopen'):
+           $HOME/Library/Frameworks, /Library/Frameworks, /Network/Library/Frameworks, /System/Library/Frameworks
+       */
+      
+      /** Loads framework by name using dl_open_lib(). This function is shortcut to easily load frameworks
+       with just it's name. If framework name is NULL or empty string, the unchanged parameter is passed.
+       Full framework name is computed only if given framework's name has length at least one character. */
+      void* dl_open_framework(const char *fw_name) {
+          if (fw_name == NULL)
+              return dl_open_lib(NULL);
+          const size_t fs = strlen(fw_name);
+          if (fs <= 0)
+              return dl_open_lib("");
+          const char *e = ".framework/";
+          const size_t es = strlen(e);
+          char *s = (char *)malloc(fs * 2 + es);
+          s = strncat(s, fw_name, fs);
+          s = strncat(s, e, es);
+          s = strncat(s, fw_name, fs);
+          void* lib = dl_open_lib(s);
+          if (s) free(s);
+          return lib;
+      }
+      
       class GEGL_EXPORT OpenGLFunctionLoader {
       protected:
           bool _triedToLoadOpenGL = false;
@@ -72,6 +115,8 @@ namespace ge{
                    (Default OpenGL path: '/System/Library/Frameworks/OpenGL.framework') */
                   const std::string frameworkName = "OpenGL";
                   this->openglLib = dl_open_framework(frameworkName.c_str());
+                  if (!this->openglLib)
+                      throw std::runtime_error("geGL::OpenGLFunctionLoader::operator() - cannot open " + frameworkName);
                   
                   this->_triedToLoadOpenGL = true;
               }
@@ -83,10 +128,10 @@ namespace ge{
                */
               
               // TODO: glGetProcAddress (SDL?)
-              return dl_symbol(this->openglLib, name);
+              return dlsym(this->openglLib, name);
           }
           ~OpenGLFunctionLoader() {
-              if(openglLib) dl_close_lib(this->openglLib);
+              if (openglLib) dl_close_lib(this->openglLib);
               this->openglLib = nullptr;
           }
       };
@@ -105,12 +150,12 @@ namespace ge{
               const std::string getProcAddressName = "glXGetProcAddress";
               if(!this->_triedToLoadOpenGL){
                   this->_triedToLoadOpenGL = true;
-                  this->openglLib = dl_open_lib(libName.c_str());
+                  this->openglLib = dlopen(libName.c_str(), RTLD_LAZY);
               }
               if(!this->_triedToLoadGetProcAddress){
                   this->_triedToLoadGetProcAddress = true;
                   if(this->openglLib)
-                      reinterpret_cast<void*&>(this->_glXGetProcAddress) = dl_symbol(this->openglLib,getProcAddressName.c_str());
+                      reinterpret_cast<void*&>(this->_glXGetProcAddress) = dlsym(this->openglLib,getProcAddressName.c_str());
                   else throw std::runtime_error("geGL::OpenGLFunctionLoader::operator() - cannot open "+libName);
               }
               if(!this->_glXGetProcAddress){
@@ -120,7 +165,7 @@ namespace ge{
               return (void*)this->_glXGetProcAddress((uint8_t const*)(name));
           }
           ~OpenGLFunctionLoader(){
-              if(openglLib)dl_close_lib(this->openglLib);
+              if(openglLib)dlclose(this->openglLib);
               this->openglLib = nullptr;
           }
       };
